Remove ports_ustring.js
