+++ racketscript-compiler/racketscript/compiler/runtime/core/ports.js
@@ -1,4 +1,5 @@
 import {PrintablePrimitive} from './printable_primitive.js';
+import * as UString from './unicode_string.js';
 
 /**
  * @abstract
@@ -139,4 +140,57 @@ export class NativeOutputStringPort extends OutputPort {
     get name() {
         return 'js-string';
     }
+}
+
+class UStringOutputPort extends OutputPort {
+    isUStringPort() {
+        return true;
+    }
+}
+
+class OutputStringPort extends UStringOutputPort {
+    constructor() {
+        super();
+        this._buffer = [];
+    }
+
+    /**
+     * @param {!UString.UString} s
+     */
+    consume(s) {
+        this._buffer.push(s);
+    }
+
+    /**
+     * @return {!UString.UString}
+     */
+    getOutputString() {
+        if (this._buffer.length === 0) {
+            return UString.makeMutable('');
+        }
+        if (this._buffer.length > 1) {
+
+            this._buffer = [UString.stringAppend(...this._buffer)];
+        }
+        return UString.copyAsMutable(this._buffer[0]);
+    }
+
+    get name() {
+        return 'string';
+    }
+}
+
+/**
+ * @return {!OutputStringPort}
+ */
+export function openOutputString() {
+    return new OutputStringPort();
+}
+
+/**
+ *
+ * @param {!OutputStringPort} outputStringPort
+ */
+export function getOutputString(outputStringPort) {
+    return outputStringPort.getOutputString();
 }
